LayoutTests/fast/forms/ios/select-options-after-optgroup-expected.txt

This test verifies that choosing options which have preceding optgroup siblings works as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS select.value is "A"
PASS select.value is "B"
PASS select.value is "2"
PASS select.value is "D"
PASS successfullyParsed is true

TEST COMPLETE
